Move thread tests into stdlib_threading.py
Merge the contents of test_threading.py and test_threading_teardown.py
into stdlib_threading.py to match the file's helper-function style; the
two source snippets are deleted. The bodies become `thread_join_ordering`
(start/join ordering) and `thread_exit_without_join` (regression test
for #7813 — no-join thread teardown must not panic).
