+++ extra_tests/snippets/stdlib_threading.py
@@ -1,6 +1,7 @@
 import multiprocessing
 import os
 import threading
+import time
 
 
 def import_in_thread(module_name):
@@ -61,6 +62,48 @@ def start_fork_process_after_thread():
     process.join(timeout=10)
     assert process.exitcode == 0, process.exitcode
 
+
+def thread_join_ordering():
+    output = []
+
+    def thread_function(name):
+        output.append((name, 0))
+        time.sleep(2.0)
+        output.append((name, 1))
+
+    output.append((0, 0))
+    x = threading.Thread(target=thread_function, args=(1,))
+    output.append((0, 1))
+    x.start()
+    output.append((0, 2))
+    x.join()
+    output.append((0, 3))
+
+    assert len(output) == 6, output
+    # CPython has [(1, 0), (0, 2)] for the middle 2, but we have [(0, 2), (1, 0)]
+    # TODO: maybe fix this, if it turns out to be a problem?
+    # assert output == [(0, 0), (0, 1), (1, 0), (0, 2), (1, 1), (0, 3)]
+
+
+def thread_exit_without_join():
+    # Regression for https://github.com/RustPython/RustPython/issues/7813:
+    # a thread started without ``.join()`` must exit cleanly even when the
+    # captured target callable drops during teardown (which can fire
+    # weakref callbacks that re-enter the VM).
+    output = []
+
+    def runner():
+        output.append("runner done")
+
+    threading.Thread(target=runner).start()
+    time.sleep(1)
+    output.append("main done")
+    assert "runner done" in output, output
+    assert "main done" in output, output
+
+
+thread_join_ordering()
+thread_exit_without_join()
 
 import_in_thread("functools")
 import_in_thread("tempfile")
