--- extra_tests/snippets/test_threading_teardown.py
@@ -1,9 +1,0 @@
-import threading
-import time
-
-def runner():
-    print("runner done")
-
-threading.Thread(target=runner).start()
-time.sleep(1)
-print("main done")
